backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibecheck-backend"
version = "0.1.0"
description = "Sentiment tracking backend for AI entities"
requires-python = ">=3.11"
authors = [
    {name = "VibeCheck Developer", email = "dev@vibecheck.example.com"}
]

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "pydantic==2.7.0",
    "sqlalchemy==2.0.35",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "httpx==0.27.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
]

[project.urls]
Homepage = "https://github.com/user/vibecheck"
Repository = "https://github.com/user/vibecheck.git"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=backend --cov-report=html"
